requirements.txt
discord.py
sqlalchemy
loguru
pytz